🎨 style: refactor code to use custom hooks for context The code has been refactored to use custom hooks for context instead of using React's useContext directly. This improves code readability and reduces the amount of boilerplate code needed to access context values. The custom hooks used are `useContexts` and `useBookmarks`.
🔨 refactor(Auth.tsx): use useContexts hook instead of ToastContext
🔨 refactor(Main.tsx): use useBookmarks hook instead of getBookmarks function
The useContexts hook is used instead of the ToastContext to improve code readability and reduce the number of imports. The useBookmarks hook is used instead of the getBookmarks function to improve code readability and reduce the number of imports. The onItemUpdate function is removed as it is no longer needed.